telegram/requirements.txt
# Telegram MCP Server dependencies
telethon
python-dotenv
mcp
fastmcp
python-dateutil